Companions have stats. Avatar DB works.
The Companions now appropriately generate stats based on the Avatar's
current stats along with particular multipliers unique to each class so
that each companion generates useful stats for their respective roles.
The Avatar database wasn't in a working condition when I last submitted
it (as expected). It has been given some general testing and should be
in relative working condition.