Attack done, SpecialAttack WIP
The normal Attack and takeDamage functions should be working as
intended. Moving on to implementing the special attacks for the avatar
now. These same special attacks will be made available to all descending
class of RPGActor which will allow for easy special ability choosing for
companions and allows the option of special abilities for RPGEnemy or
w/e they will be called.