Events are now Completable! (minor problem though)
There is a strange bug where only 1 event is schedulable because of
"conflicts."  I tried to debug to find the problem but it all seems to
be fine and, well when you debug there IS no problem, it actually lets
you add it.  But at regular runtime it does not.