RPG FINAL
Lots of balancing has occurred in the RPG portion of the game along with
properly displaying the game over/victory screens and updating your
money/battles won accordingly. There is floating combat text which is
facilitated with Toasts and they don't overlap terribly like they used
to. I also fixed a lot of our usage of Random number generators (they
were producing negative values when they shouldn't have been). This
fixed a lot of the issues with the music as well as RPG elements.